// src/components/organisms/OldUpdateKeyword/interfaces.ts
export interface IKeywordEligibilityKeywordShift {
	from: any;
	to: any;
}

export interface IKeywordEligibility {
	name: string;
	start_period: any;
	end_period: any;
	keyword_type: string;
	point_type: string;
	poin_value: string;
	poin_redeemed: number;
	channel_validation: boolean;
	channel_validation_list: string[];
	program_id: string;
	eligibility_locations: boolean;
	location_type: string;
	locations: string[];
	program_title_expose: string;
	program_experience: string[];
	program_bersubsidi: boolean;
	merchant: string;
	merchandise_keyword: boolean;
	keyword_schedule: string;
	total_budget: number;
	customer_value: number;
	multiwhitelist: boolean;
	multiwhitelist_program: string;
	enable_sms_masking: boolean;
	sms_masking: string;
	timezone: string;
	for_new_redeemer: boolean;
	max_mode: string;
	max_redeem_counter: number;
	segmentation_customer_tier: string[];
	segmentation_customer_los_operator: string;
	segmentation_customer_los: number;
	segmentation_customer_los_min: number;
	segmentation_customer_los_max: number;
	segmentation_customer_type: string;
	segmentation_customer_most_redeem: string[];
	segmentation_customer_brand: string[];
	segmentation_customer_prepaid_registration: boolean;
	segmentation_customer_kyc_completeness: boolean;
	segmentation_customer_poin_balance_operator: string;
	segmentation_customer_poin_balance: number;
	segmentation_customer_poin_balance_min: number;
	segmentation_customer_poin_balance_max: number;
	segmentation_customer_preference: string;
	segmentation_customer_arpu_operator: string;
	segmentation_customer_arpu: number;
	segmentation_customer_arpu_min: number;
	segmentation_customer_arpu_max: number;
	segmentation_customer_preferences_bcp: string;
	file: string;
	segmentation_employee_numbers: boolean;
	keyword_shift: IKeywordEligibilityKeywordShift[];
}
// ...
export interface IUpdateKeyword {
	eligibility: IKeywordEligibility;
	bonus: any[];
	notification: any[];
}